+++ app/javascript/vue/tasks/clipboard/app.vue
@@ -1,145 +1,127 @@
 <template>
   <div
     id="vue-clipboard-app"
-    class="full_width">
+    class="full_width"
+  >
     <ul class="slide-panel-category-content">
       <li
-        v-for="(clip, index) in clipboard"
+        v-for="(_, index) in clipboard"
         :key="index"
-        class="slide-panel-category-item">
-        <div class="full_width padding-large-right">
-          <p>Shortcut: <b>{{ pasteKeys.join(' + ') }} + {{ index }}</b></p>
+        class="slide-panel-category-item"
+      >
+        <div class="full_width padding-large-right capitalize">
+          <p>Shortcut: <b>{{ actionKey }} + {{ keyPaste }} + {{ index }}</b></p>
           <div class="middle">
             <textarea
               class="full_width"
               rows="5"
               v-model="clipboard[index]"
-              @blur="saveClipboard"/>
+              @blur="saveClipboard"
+            />
           </div>
         </div>
       </li>
     </ul>
     <p class="slide-panel-category-content">
-      Use <b>{{ actionKey }} + {{ keyCopy }} + Number</b> to copy a text to the clipboard box
+      Use <b class="capitalize">{{ actionKey }} + {{ keyCopy }} + Number</b> to copy a text to the clipboard box
     </p>
   </div>
 </template>
-<script>
-
+<script setup>
+import { ref, onBeforeMount, onBeforeUnmount } from 'vue'
 import { ProjectMember } from 'routes/endpoints'
+import platformKey from 'helpers/getPlatformKey'
+import useHotkey from 'vue3-hotkey'
 
-export default {
-  name: 'ClipboardApp',
+const clipboard = ref({
+  1: '',
+  2: '',
+  3: '',
+  4: '',
+  5: ''
+})
 
-  computed: {
-    actionKey () {
-      return navigator.platform.indexOf('Mac') > -1
-        ? 'Control'
-        : 'Alt'
-    },
+const actionKey = platformKey()
+const keyCopy = 'c'
+const keyPaste = 'v'
 
-    isLinux () {
-      return navigator.platform.indexOf('Linux') > -1
-    },
+function generateShortcuts () {
+  const hotkeys = []
 
-    pasteKeys () {
-      return this.isLinux
-        ? [this.actionKey, 'Control']
-        : [this.actionKey]
-    }
-  },
-
-  data () {
-    return {
-      clipboard: {
-        1: '',
-        2: '',
-        3: '',
-        4: '',
-        5: ''
-      },
-      keys: [],
-      keyCopy: 'C'
-    }
-  },
-
-  created () {
-    document.addEventListener('turbolinks:load', () => {
-      window.removeEventListener('keydown', this.keyPressed)
-      window.removeEventListener('keyup', this.removeKey)
+  Object.keys(clipboard.value).forEach(slot => {
+    hotkeys.push({
+      keys: [actionKey, keyPaste, slot],
+      preventDefault: true,
+      handler () {
+        pasteClipboard(slot)
+      }
     })
 
-    window.addEventListener('keydown', this.keyPressed)
-    window.addEventListener('keyup', this.removeKey)
-
-    ProjectMember.clipboard().then(response => {
-      Object.assign(this.clipboard, response.body.clipboard)
+    hotkeys.push({
+      keys: [actionKey, keyCopy, slot],
+      preventDefault: true,
+      handler () {
+        setClipboard(slot)
+      }
     })
-  },
+  })
 
-  methods: {
-    isInput () {
-      return document.activeElement.tagName === 'INPUT' ||
-          document.activeElement.tagName === 'TEXTAREA'
-    },
+  return hotkeys
+}
 
-    keyPressed (event) {
-      const { code, key } = event
-      const keyPressed = String(Object.keys(this.clipboard).findIndex(keyCode => `Digit${keyCode}` === code) + 1)
-      const isClipboardKey = Object.keys(this.clipboard).includes(key)
-      const iskeyCopyPressed = !!this.keys.find(key => key.toUpperCase() === this.keyCopy)
+const hotkeys = ref(generateShortcuts())
 
-      this.addKey(isClipboardKey ? keyPressed : key)
+TW.workbench.keyboard.createLegend(`${platformKey()}+f`, 'Search', 'Filter sources')
+TW.workbench.keyboard.createLegend(`${platformKey()}+r`, 'Reset task', 'Filter sources')
 
-      if ((this.keys.includes(this.actionKey) && event.getModifierState(this.actionKey)) && isClipboardKey) {
-        if (iskeyCopyPressed) {
-          this.setClipboard(key)
-        } else if (this.pasteKeys.every(key => this.keys.includes(key))) {
-          this.pasteClipboard(key)
-        }
-        event.preventDefault()
-      }
-    },
+const stop = useHotkey(hotkeys.value)
 
-    pasteClipboard (clipboardIndex) {
-      if (this.isInput() && this.clipboard[clipboardIndex]) {
-        const position = document.activeElement.selectionStart
-        const text = document.activeElement.value
-        document.activeElement.value = text.substr(0, position) + this.clipboard[clipboardIndex] + text.substr(position)
-        document.activeElement.dispatchEvent(new CustomEvent('input'))
-      }
-    },
+onBeforeMount(() => {
+  ProjectMember.clipboard().then(response => {
+    Object.assign(clipboard.value, response.body.clipboard)
+  })
+})
 
-    saveClipboard () {
-      ProjectMember.updateClipboard(this.clipboard).then(response => {
-        this.clipboard = response.body.clipboard
-      })
-    },
+onBeforeUnmount(() => {
+  stop.forEach(hotkey => hotkey())
+})
 
-    setClipboard (index) {
-      const textSelected = this.isInput()
-        ? document.activeElement.value
-        : window.getSelection().toString()
+function isInput () {
+  return document.activeElement.tagName === 'INPUT' ||
+      document.activeElement.tagName === 'TEXTAREA'
+}
 
-      if (textSelected.length > 0) {
-        this.clipboard[index] = textSelected
-        this.saveClipboard()
-      }
-    },
+function pasteClipboard (clipboardIndex) {
+  if (isInput() && clipboard.value[clipboardIndex]) {
+    const position = document.activeElement.selectionStart
+    const text = document.activeElement.value
 
-    addKey (key) {
-      if (!this.keys.includes(key)) {
-        this.keys.push(key)
-      }
-    },
-
-    removeKey ({ key }) {
-      const position = this.keys.findIndex(keyStore => keyStore === key)
-
-      if (position > -1) {
-        this.keys.splice(position, 1)
-      }
-    }
+    document.activeElement.value = text.substr(0, position) + clipboard.value[clipboardIndex] + text.substr(position)
+    document.activeElement.dispatchEvent(new CustomEvent('input'))
   }
+}
+
+function saveClipboard () {
+  ProjectMember.updateClipboard(clipboard.value).then(response => {
+    clipboard.value = response.body.clipboard
+  })
+}
+
+function setClipboard (index) {
+  const textSelected = isInput()
+    ? document.activeElement.value
+    : window.getSelection().toString()
+
+  if (textSelected.length > 0) {
+    clipboard.value[index] = textSelected
+    saveClipboard()
+  }
+}
+
+</script>
+
+<script>
+export default {
+  name: 'ClipboardApp'
 }
 </script>